Fix handling with full access / blocked devices
