+++ supervisor/hardware/module.py
@@ -10,6 +10,7 @@ from supervisor.hardware.const import UdevSubsystem
 from ..coresys import CoreSys, CoreSysAttributes
 from ..exceptions import HardwareNotFound
 from .data import Device
+from .disk import HwDisk
 from .helper import HwHelper
 from .monitor import HwMonitor
 from .policy import HwPolicy
@@ -29,6 +30,7 @@ class HardwareManager(CoreSysAttributes):
         self._montior: HwMonitor = HwMonitor(coresys)
         self._helper: HwHelper = HwHelper(coresys)
         self._policy: HwPolicy = HwPolicy(coresys)
+        self._disk: HwDisk = HwDisk(coresys)
 
     @property
     def monitor(self) -> HwMonitor:
@@ -44,6 +46,11 @@ class HardwareManager(CoreSysAttributes):
     def policy(self) -> HwPolicy:
         """Return Hardware policy instance."""
         return self._policy
+
+    @property
+    def disk(self) -> HwDisk:
+        """Return Hardware disk instance."""
+        return self._disk
 
     @property
     def devices(self) -> List[Device]:
